+++ apps/api/src/app/controllers/app.controller.spec.ts
@@ -1,7 +1,7 @@
 import { Test, TestingModule } from '@nestjs/testing';
 
 import { AppController } from './app.controller';
-import { AppService } from './app.service';
+import { AppService } from '../api-services/app.service';
 
 describe('AppController', () => {
   let app: TestingModule;
