+++ apps/api/src/app/controllers/app.controller.ts
@@ -1,6 +1,6 @@
 import { Controller, Get } from '@nestjs/common';
 
-import { AppService } from './app.service';
+import { AppService } from '../api-services/app.service';
 
 @Controller()
 export class AppController {
