+++ apps/api/src/app/controllers/players/players.controller.ts
@@ -1,5 +1,5 @@
 import { Controller, Get } from '@nestjs/common';
-import { PlayersService } from './players.service';
+import { PlayersService } from '../../api-services/players/players.service';
 
 @Controller('players')
 export class PlayersController {
