--- libs/players/README.md
@@ -1,7 +1,0 @@
-# players
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test players` to execute the unit tests via [Jest](https://jestjs.io).
