--- libs/players/jest.config.ts
@@ -1,11 +1,0 @@
-/* eslint-disable */
-export default {
-  displayName: 'players',
-  preset: '../../jest.preset.js',
-  testEnvironment: 'node',
-  transform: {
-    '^.+\\.[tj]s$': ['ts-jest', { tsconfig: '<rootDir>/tsconfig.spec.json' }],
-  },
-  moduleFileExtensions: ['ts', 'js', 'html'],
-  coverageDirectory: '../../coverage/libs/players',
-};
