--- libs/players/src/index.ts
@@ -1,3 +1,0 @@
-export * from './lib/players.controller';
-export * from './lib/players.service';
-export * from './lib/players.module';
