--- libs/players/src/lib/players.module.ts
@@ -1,10 +1,0 @@
-import { Module } from '@nestjs/common';
-import { PlayersController } from './players.controller';
-import { PlayersService } from './players.service';
-
-@Module({
-  controllers: [PlayersController],
-  providers: [PlayersService],
-  exports: [PlayersService],
-})
-export class PlayersModule {}
